update to pre-build validation code
