--- packages/next/compiled/ajv-keywords/index.js
@@ -1,1 +1,0 @@
-(()=>{"use strict";var e={409:(e,t)=>{Object.defineProperty(t,"__esModule",{value:true});function getRangeDef(e){return()=>({keyword:e,type:"number",schemaType:"array",macro:function([t,r]){validateRangeSchema(t,r);return e==="range"?{minimum:t,maximum:r}:{exclusiveMinimum:t,exclusiveMaximum:r}},metaSchema:{type:"array",minItems:2,maxItems:2,items:{type:"number"}}});function validateRangeSchema(t,r){if(t>r||e==="exclusiveRange"&&t===r){throw new Error("There are no numbers in range")}}}t["default"]=getRangeDef},666:(e,t)=>{Object.defineProperty(t,"__esModule",{value:true});function getRequiredDef(e){return()=>({keyword:e,type:"object",schemaType:"array",macro(t){if(t.length===0)return true;if(t.length===1)return{required:t};const r=e==="anyRequired"?"anyOf":"oneOf";return{[r]:t.map((e=>({required:[e]})))}},metaSchema:{type:"array",items:{type:"string"}}})}t["default"]=getRequiredDef},775:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});t.usePattern=t.metaSchemaRef=void 0;const n=r(794);const s="http://json-schema.org/schema";function metaSchemaRef({defaultMeta:e}={}){return e===false?{}:{$ref:e||s}}t.metaSchemaRef=metaSchemaRef;function usePattern({gen:e,it:{opts:t}},r,s=(t.unicodeRegExp?"u":"")){const o=new RegExp(r,s);return e.scopeValue("pattern",{key:o.toString(),ref:o,code:(0,n._)`new RegExp(${r}, ${s})`})}t.usePattern=usePattern},830:(e,t)=>{Object.defineProperty(t,"__esModule",{value:true});function getDef(){return{keyword:"allRequired",type:"object",schemaType:"boolean",macro(e,t){if(!e)return true;const r=Object.keys(t.properties);if(r.length===0)return true;return{required:r}},dependencies:["properties"]}}t["default"]=getDef;e.exports=getDef},550:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(666));const o=(0,s.default)("anyRequired");t["default"]=o;e.exports=o},774:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});const n=r(775);function getDef(e){return{keyword:"deepProperties",type:"object",schemaType:"object",macro:function(e){const t=[];for(const r in e)t.push(getSchema(r,e[r]));return{allOf:t}},metaSchema:{type:"object",propertyNames:{type:"string",format:"json-pointer"},additionalProperties:(0,n.metaSchemaRef)(e)}}}t["default"]=getDef;function getSchema(e,t){const r=e.split("/");const n={};let s=n;for(let e=1;e<r.length;e++){let n=r[e];const o=e===r.length-1;n=unescapeJsonPointer(n);const i=s.properties={};let a;if(/[0-9]+/.test(n)){let e=+n;a=s.items=[];s.type=["object","array"];while(e--)a.push({})}else{s.type="object"}s=o?t:{};i[n]=s;if(a)a.push(s)}return n}function unescapeJsonPointer(e){return e.replace(/~1/g,"/").replace(/~0/g,"~")}e.exports=getDef},464:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});const n=r(794);function getDef(){return{keyword:"deepRequired",type:"object",schemaType:"array",code(e){const{schema:t,data:r}=e;const s=t.map((e=>(0,n._)`(${getData(e)}) === undefined`));e.fail((0,n.or)(...s));function getData(e){if(e==="")throw new Error("empty JSON pointer not allowed");const t=e.split("/");let s=r;const o=t.map(((e,t)=>t?s=(0,n._)`${s}${(0,n.getProperty)(unescapeJPSegment(e))}`:s));return(0,n.and)(...o)}},metaSchema:{type:"array",items:{type:"string",format:"json-pointer"}}}}t["default"]=getDef;function unescapeJPSegment(e){return e.replace(/~1/g,"/").replace(/~0/g,"~")}e.exports=getDef},407:(e,t)=>{Object.defineProperty(t,"__esModule",{value:true});const r={};const n={timestamp:()=>()=>Date.now(),datetime:()=>()=>(new Date).toISOString(),date:()=>()=>(new Date).toISOString().slice(0,10),time:()=>()=>(new Date).toISOString().slice(11),random:()=>()=>Math.random(),randomint:e=>{var t;const r=(t=e===null||e===void 0?void 0:e.max)!==null&&t!==void 0?t:2;return()=>Math.floor(Math.random()*r)},seq:e=>{var t;const n=(t=e===null||e===void 0?void 0:e.name)!==null&&t!==void 0?t:"";r[n]||(r[n]=0);return()=>r[n]++}};const s=Object.assign(_getDef,{DEFAULTS:n});function _getDef(){return{keyword:"dynamicDefaults",type:"object",schemaType:["string","object"],modifying:true,valid:true,compile(e,t,r){if(!r.opts.useDefaults||r.compositeRule)return()=>true;const n={};for(const t in e)n[t]=getDefault(e[t]);const s=r.opts.useDefaults==="empty";return t=>{for(const r in e){if(t[r]===undefined||s&&(t[r]===null||t[r]==="")){t[r]=n[r]()}}return true}},metaSchema:{type:"object",additionalProperties:{anyOf:[{type:"string"},{type:"object",additionalProperties:false,required:["func","args"],properties:{func:{type:"string"},args:{type:"object"}}}]}}}}function getDefault(e){return typeof e=="object"?getObjDefault(e):getStrDefault(e)}function getObjDefault({func:e,args:t}){const r=n[e];assertDefined(e,r);return r(t)}function getStrDefault(e=""){const t=n[e];assertDefined(e,t);return t()}function assertDefined(e,t){if(!t)throw new Error(`invalid "dynamicDefaults" keyword property value: ${e}`)}t["default"]=s;e.exports=s},4:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(409));const o=(0,s.default)("exclusiveRange");t["default"]=o;e.exports=o},372:(e,t)=>{Object.defineProperty(t,"__esModule",{value:true});const r={Object:Object,Array:Array,Function:Function,Number:Number,String:String,Date:Date,RegExp:RegExp};if(typeof Buffer!="undefined")r.Buffer=Buffer;if(typeof Promise!="undefined")r.Promise=Promise;const n=Object.assign(_getDef,{CONSTRUCTORS:r});function _getDef(){return{keyword:"instanceof",schemaType:["string","array"],compile(e){if(typeof e=="string"){const t=getConstructor(e);return e=>e instanceof t}if(Array.isArray(e)){const t=e.map(getConstructor);return e=>{for(const r of t){if(e instanceof r)return true}return false}}throw new Error("ajv implementation error")},metaSchema:{anyOf:[{type:"string"},{type:"array",items:{type:"string"}}]}}}function getConstructor(e){const t=r[e];if(t)return t;throw new Error(`invalid "instanceof" keyword value ${e}`)}t["default"]=n;e.exports=n},762:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(666));const o=(0,s.default)("oneRequired");t["default"]=o;e.exports=o},816:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});const n=r(794);const s=r(775);const o={message:({params:{missingPattern:e}})=>(0,n.str)`should have property matching pattern '${e}'`,params:({params:{missingPattern:e}})=>(0,n._)`{missingPattern: ${e}}`};function getDef(){return{keyword:"patternRequired",type:"object",schemaType:"array",error:o,code(e){const{gen:t,schema:r,data:o}=e;if(r.length===0)return;const i=t.let("valid",true);for(const e of r)validateProperties(e);function validateProperties(r){const a=t.let("matched",false);t.forIn("key",o,(o=>{t.assign(a,(0,n._)`${(0,s.usePattern)(e,r)}.test(${o})`);t.if(a,(()=>t.break()))}));e.setParams({missingPattern:r});t.assign(i,(0,n.and)(i,a));e.pass(i)}},metaSchema:{type:"array",items:{type:"string",format:"regex"},uniqueItems:true}}}t["default"]=getDef;e.exports=getDef},583:(e,t)=>{Object.defineProperty(t,"__esModule",{value:true});function getDef(){return{keyword:"prohibited",type:"object",schemaType:"array",macro:function(e){if(e.length===0)return true;if(e.length===1)return{not:{required:e}};return{not:{anyOf:e.map((e=>({required:[e]})))}}},metaSchema:{type:"array",items:{type:"string"}}}}t["default"]=getDef;e.exports=getDef},851:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(409));const o=(0,s.default)("range");t["default"]=o;e.exports=o},545:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});const n=r(794);const s=r(775);const o={type:"object",properties:{pattern:{type:"string"},flags:{type:"string",nullable:true}},required:["pattern"],additionalProperties:false};const i=/^\/(.*)\/([gimuy]*)$/;function getDef(){return{keyword:"regexp",type:"string",schemaType:["string","object"],code(e){const{data:t,schema:r}=e;const o=getRegExp(r);e.pass((0,n._)`${o}.test(${t})`);function getRegExp(t){if(typeof t=="object")return(0,s.usePattern)(e,t.pattern,t.flags);const r=i.exec(t);if(r)return(0,s.usePattern)(e,r[1],r[2]);throw new Error("cannot parse string into RegExp")}},metaSchema:{anyOf:[{type:"string"},o]}}}t["default"]=getDef;e.exports=getDef},637:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});const n=r(794);const s=r(775);const o={message:({params:{schemaProp:e}})=>e?(0,n.str)`should match case "${e}" schema`:(0,n.str)`should match default case schema`,params:({params:{schemaProp:e}})=>e?(0,n._)`{failingCase: ${e}}`:(0,n._)`{failingDefault: true}`};function getDef(e){const t=(0,s.metaSchemaRef)(e);return[{keyword:"select",schemaType:["string","number","boolean","null"],$data:true,error:o,dependencies:["selectCases"],code(e){const{gen:t,schemaCode:r,parentSchema:s}=e;e.block$data(n.nil,(()=>{const o=t.let("valid",true);const i=t.name("_valid");const a=t.const("value",(0,n._)`${r} === null ? "null" : ${r}`);t.if(false);for(const r in s.selectCases){e.setParams({schemaProp:r});t.elseIf((0,n._)`"" + ${a} == ${r}`);const s=e.subschema({keyword:"selectCases",schemaProp:r},i);e.mergeEvaluated(s,n.Name);t.assign(o,i)}t.else();if(s.selectDefault!==undefined){e.setParams({schemaProp:undefined});const r=e.subschema({keyword:"selectDefault"},i);e.mergeEvaluated(r,n.Name);t.assign(o,i)}t.endIf();e.pass(o)}))}},{keyword:"selectCases",dependencies:["select"],metaSchema:{type:"object",additionalProperties:t}},{keyword:"selectDefault",dependencies:["select","selectCases"],metaSchema:t}]}t["default"]=getDef;e.exports=getDef},542:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});const n=r(794);const s={trimStart:e=>e.trimStart(),trimEnd:e=>e.trimEnd(),trimLeft:e=>e.trimStart(),trimRight:e=>e.trimEnd(),trim:e=>e.trim(),toLowerCase:e=>e.toLowerCase(),toUpperCase:e=>e.toUpperCase(),toEnumCase:(e,t)=>(t===null||t===void 0?void 0:t.hash[configKey(e)])||e};const o=Object.assign(_getDef,{transform:s});function _getDef(){return{keyword:"transform",schemaType:"array",before:"enum",code(e){const{gen:t,data:r,schema:o,parentSchema:i,it:a}=e;const{parentData:u,parentDataProperty:c}=a;const d=o;if(!d.length)return;let f;if(d.includes("toEnumCase")){const e=getEnumCaseCfg(i);f=t.scopeValue("obj",{ref:e,code:(0,n.stringify)(e)})}t.if((0,n._)`typeof ${r} == "string" && ${u} !== undefined`,(()=>{t.assign(r,transformExpr(d.slice()));t.assign((0,n._)`${u}[${c}]`,r)}));function transformExpr(e){if(!e.length)return r;const o=e.pop();if(!(o in s))throw new Error(`transform: unknown transformation ${o}`);const i=t.scopeValue("func",{ref:s[o],code:(0,n._)`require("ajv-keywords/dist/definitions/transform").transform${(0,n.getProperty)(o)}`});const a=transformExpr(e);return f&&o==="toEnumCase"?(0,n._)`${i}(${a}, ${f})`:(0,n._)`${i}(${a})`}},metaSchema:{type:"array",items:{type:"string",enum:Object.keys(s)}}}}function getEnumCaseCfg(e){const t={hash:{}};if(!e.enum)throw new Error('transform: "toEnumCase" requires "enum"');for(const r of e.enum){if(typeof r!=="string")continue;const e=configKey(r);if(t.hash[e]){throw new Error('transform: "toEnumCase" requires all lowercased "enum" values to be unique')}t.hash[e]=r}return t}function configKey(e){return e.toLowerCase()}t["default"]=o;e.exports=o},227:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});const n=r(794);const s=["undefined","string","number","object","function","boolean","symbol"];function getDef(){return{keyword:"typeof",schemaType:["string","array"],code(e){const{data:t,schema:r,schemaValue:s}=e;e.fail(typeof r=="string"?(0,n._)`typeof ${t} != ${r}`:(0,n._)`${s}.indexOf(typeof ${t}) < 0`)},metaSchema:{anyOf:[{type:"string",enum:s},{type:"array",items:{type:"string",enum:s}}]}}}t["default"]=getDef;e.exports=getDef},639:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});const n=r(914);const s=["number","integer","string","boolean","null"];function getDef(){return{keyword:"uniqueItemProperties",type:"array",schemaType:"array",compile(e,t){const r=getScalarKeys(e,t);return t=>{if(t.length<=1)return true;for(let s=0;s<e.length;s++){const o=e[s];if(r[s]){const e={};for(const r of t){if(!r||typeof r!="object")continue;let t=r[o];if(t&&typeof t=="object")continue;if(typeof t=="string")t='"'+t;if(e[t])return false;e[t]=true}}else{for(let e=t.length;e--;){const r=t[e];if(!r||typeof r!="object")continue;for(let s=e;s--;){const e=t[s];if(e&&typeof e=="object"&&n(r[o],e[o]))return false}}}}return true}},metaSchema:{type:"array",items:{type:"string"}}}}t["default"]=getDef;function getScalarKeys(e,t){return e.map((e=>{var r,n,o;const i=(o=(n=(r=t.items)===null||r===void 0?void 0:r.properties)===null||n===void 0?void 0:n[e])===null||o===void 0?void 0:o.type;return Array.isArray(i)?!i.includes("object")&&!i.includes("array"):s.includes(i)}))}e.exports=getDef},523:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(149));const ajvKeywords=(e,t)=>{if(Array.isArray(t)){for(const r of t)get(r)(e);return e}if(t){get(t)(e);return e}for(t in s.default)get(t)(e);return e};ajvKeywords.get=get;function get(e){const t=s.default[e];if(!t)throw new Error("Unknown keyword "+e);return t}t["default"]=ajvKeywords;e.exports=ajvKeywords;e.exports["default"]=ajvKeywords},63:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(830));const allRequired=e=>e.addKeyword((0,s.default)());t["default"]=allRequired;e.exports=allRequired},890:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(550));const anyRequired=e=>e.addKeyword((0,s.default)());t["default"]=anyRequired;e.exports=anyRequired},952:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(774));const deepProperties=(e,t)=>e.addKeyword((0,s.default)(t));t["default"]=deepProperties;e.exports=deepProperties},97:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(464));const deepRequired=e=>e.addKeyword((0,s.default)());t["default"]=deepRequired;e.exports=deepRequired},949:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(407));const dynamicDefaults=e=>e.addKeyword((0,s.default)());t["default"]=dynamicDefaults;e.exports=dynamicDefaults},738:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(4));const exclusiveRange=e=>e.addKeyword((0,s.default)());t["default"]=exclusiveRange;e.exports=exclusiveRange},149:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(432));const o=n(r(825));const i=n(r(716));const a=n(r(738));const u=n(r(769));const c=n(r(928));const d=n(r(165));const f=n(r(63));const l=n(r(890));const p=n(r(257));const h=n(r(232));const m=n(r(371));const g=n(r(952));const y=n(r(97));const v=n(r(949));const w=n(r(863));const b={typeof:s.default,instanceof:o.default,range:i.default,exclusiveRange:a.default,regexp:u.default,transform:c.default,uniqueItemProperties:d.default,allRequired:f.default,anyRequired:l.default,oneRequired:p.default,patternRequired:h.default,prohibited:m.default,deepProperties:g.default,deepRequired:y.default,dynamicDefaults:v.default,select:w.default};t["default"]=b;e.exports=b},825:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(372));const instanceofPlugin=e=>e.addKeyword((0,s.default)());t["default"]=instanceofPlugin;e.exports=instanceofPlugin},257:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(762));const oneRequired=e=>e.addKeyword((0,s.default)());t["default"]=oneRequired;e.exports=oneRequired},232:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(816));const patternRequired=e=>e.addKeyword((0,s.default)());t["default"]=patternRequired;e.exports=patternRequired},371:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(583));const prohibited=e=>e.addKeyword((0,s.default)());t["default"]=prohibited;e.exports=prohibited},716:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(851));const range=e=>e.addKeyword((0,s.default)());t["default"]=range;e.exports=range},769:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(545));const regexp=e=>e.addKeyword((0,s.default)());t["default"]=regexp;e.exports=regexp},863:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(637));const select=(e,t)=>{(0,s.default)(t).forEach((t=>e.addKeyword(t)));return e};t["default"]=select;e.exports=select},928:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(542));const transform=e=>e.addKeyword((0,s.default)());t["default"]=transform;e.exports=transform},432:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(227));const typeofPlugin=e=>e.addKeyword((0,s.default)());t["default"]=typeofPlugin;e.exports=typeofPlugin},165:function(e,t,r){var n=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:true});const s=n(r(639));const uniqueItemProperties=e=>e.addKeyword((0,s.default)());t["default"]=uniqueItemProperties;e.exports=uniqueItemProperties},621:(e,t)=>{Object.defineProperty(t,"__esModule",{value:true});t.regexpCode=t.getEsmExportName=t.getProperty=t.safeStringify=t.stringify=t.strConcat=t.addCodeArg=t.str=t._=t.nil=t._Code=t.Name=t.IDENTIFIER=t._CodeOrName=void 0;class _CodeOrName{}t._CodeOrName=_CodeOrName;t.IDENTIFIER=/^[a-z$_][a-z$_0-9]*$/i;class Name extends _CodeOrName{constructor(e){super();if(!t.IDENTIFIER.test(e))throw new Error("CodeGen: name must be a valid identifier");this.str=e}toString(){return this.str}emptyStr(){return false}get names(){return{[this.str]:1}}}t.Name=Name;class _Code extends _CodeOrName{constructor(e){super();this._items=typeof e==="string"?[e]:e}toString(){return this.str}emptyStr(){if(this._items.length>1)return false;const e=this._items[0];return e===""||e==='""'}get str(){var e;return(e=this._str)!==null&&e!==void 0?e:this._str=this._items.reduce(((e,t)=>`${e}${t}`),"")}get names(){var e;return(e=this._names)!==null&&e!==void 0?e:this._names=this._items.reduce(((e,t)=>{if(t instanceof Name)e[t.str]=(e[t.str]||0)+1;return e}),{})}}t._Code=_Code;t.nil=new _Code("");function _(e,...t){const r=[e[0]];let n=0;while(n<t.length){addCodeArg(r,t[n]);r.push(e[++n])}return new _Code(r)}t._=_;const r=new _Code("+");function str(e,...t){const n=[safeStringify(e[0])];let s=0;while(s<t.length){n.push(r);addCodeArg(n,t[s]);n.push(r,safeStringify(e[++s]))}optimize(n);return new _Code(n)}t.str=str;function addCodeArg(e,t){if(t instanceof _Code)e.push(...t._items);else if(t instanceof Name)e.push(t);else e.push(interpolate(t))}t.addCodeArg=addCodeArg;function optimize(e){let t=1;while(t<e.length-1){if(e[t]===r){const r=mergeExprItems(e[t-1],e[t+1]);if(r!==undefined){e.splice(t-1,3,r);continue}e[t++]="+"}t++}}function mergeExprItems(e,t){if(t==='""')return e;if(e==='""')return t;if(typeof e=="string"){if(t instanceof Name||e[e.length-1]!=='"')return;if(typeof t!="string")return`${e.slice(0,-1)}${t}"`;if(t[0]==='"')return e.slice(0,-1)+t.slice(1);return}if(typeof t=="string"&&t[0]==='"'&&!(e instanceof Name))return`"${e}${t.slice(1)}`;return}function strConcat(e,t){return t.emptyStr()?e:e.emptyStr()?t:str`${e}${t}`}t.strConcat=strConcat;function interpolate(e){return typeof e=="number"||typeof e=="boolean"||e===null?e:safeStringify(Array.isArray(e)?e.join(","):e)}function stringify(e){return new _Code(safeStringify(e))}t.stringify=stringify;function safeStringify(e){return JSON.stringify(e).replace(/\u2028/g,"\\u2028").replace(/\u2029/g,"\\u2029")}t.safeStringify=safeStringify;function getProperty(e){return typeof e=="string"&&t.IDENTIFIER.test(e)?new _Code(`.${e}`):_`[${e}]`}t.getProperty=getProperty;function getEsmExportName(e){if(typeof e=="string"&&t.IDENTIFIER.test(e)){return new _Code(`${e}`)}throw new Error(`CodeGen: invalid export name: ${e}, use explicit $id name mapping`)}t.getEsmExportName=getEsmExportName;function regexpCode(e){return new _Code(e.toString())}t.regexpCode=regexpCode},794:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});t.or=t.and=t.not=t.CodeGen=t.operators=t.varKinds=t.ValueScopeName=t.ValueScope=t.Scope=t.Name=t.regexpCode=t.stringify=t.getProperty=t.nil=t.strConcat=t.str=t._=void 0;const n=r(621);const s=r(997);var o=r(621);Object.defineProperty(t,"_",{enumerable:true,get:function(){return o._}});Object.defineProperty(t,"str",{enumerable:true,get:function(){return o.str}});Object.defineProperty(t,"strConcat",{enumerable:true,get:function(){return o.strConcat}});Object.defineProperty(t,"nil",{enumerable:true,get:function(){return o.nil}});Object.defineProperty(t,"getProperty",{enumerable:true,get:function(){return o.getProperty}});Object.defineProperty(t,"stringify",{enumerable:true,get:function(){return o.stringify}});Object.defineProperty(t,"regexpCode",{enumerable:true,get:function(){return o.regexpCode}});Object.defineProperty(t,"Name",{enumerable:true,get:function(){return o.Name}});var i=r(997);Object.defineProperty(t,"Scope",{enumerable:true,get:function(){return i.Scope}});Object.defineProperty(t,"ValueScope",{enumerable:true,get:function(){return i.ValueScope}});Object.defineProperty(t,"ValueScopeName",{enumerable:true,get:function(){return i.ValueScopeName}});Object.defineProperty(t,"varKinds",{enumerable:true,get:function(){return i.varKinds}});t.operators={GT:new n._Code(">"),GTE:new n._Code(">="),LT:new n._Code("<"),LTE:new n._Code("<="),EQ:new n._Code("==="),NEQ:new n._Code("!=="),NOT:new n._Code("!"),OR:new n._Code("||"),AND:new n._Code("&&"),ADD:new n._Code("+")};class Node{optimizeNodes(){return this}optimizeNames(e,t){return this}}class Def extends Node{constructor(e,t,r){super();this.varKind=e;this.name=t;this.rhs=r}render({es5:e,_n:t}){const r=e?s.varKinds.var:this.varKind;const n=this.rhs===undefined?"":` = ${this.rhs}`;return`${r} ${this.name}${n};`+t}optimizeNames(e,t){if(!e[this.name.str])return;if(this.rhs)this.rhs=optimizeExpr(this.rhs,e,t);return this}get names(){return this.rhs instanceof n._CodeOrName?this.rhs.names:{}}}class Assign extends Node{constructor(e,t,r){super();this.lhs=e;this.rhs=t;this.sideEffects=r}render({_n:e}){return`${this.lhs} = ${this.rhs};`+e}optimizeNames(e,t){if(this.lhs instanceof n.Name&&!e[this.lhs.str]&&!this.sideEffects)return;this.rhs=optimizeExpr(this.rhs,e,t);return this}get names(){const e=this.lhs instanceof n.Name?{}:{...this.lhs.names};return addExprNames(e,this.rhs)}}class AssignOp extends Assign{constructor(e,t,r,n){super(e,r,n);this.op=t}render({_n:e}){return`${this.lhs} ${this.op}= ${this.rhs};`+e}}class Label extends Node{constructor(e){super();this.label=e;this.names={}}render({_n:e}){return`${this.label}:`+e}}class Break extends Node{constructor(e){super();this.label=e;this.names={}}render({_n:e}){const t=this.label?` ${this.label}`:"";return`break${t};`+e}}class Throw extends Node{constructor(e){super();this.error=e}render({_n:e}){return`throw ${this.error};`+e}get names(){return this.error.names}}class AnyCode extends Node{constructor(e){super();this.code=e}render({_n:e}){return`${this.code};`+e}optimizeNodes(){return`${this.code}`?this:undefined}optimizeNames(e,t){this.code=optimizeExpr(this.code,e,t);return this}get names(){return this.code instanceof n._CodeOrName?this.code.names:{}}}class ParentNode extends Node{constructor(e=[]){super();this.nodes=e}render(e){return this.nodes.reduce(((t,r)=>t+r.render(e)),"")}optimizeNodes(){const{nodes:e}=this;let t=e.length;while(t--){const r=e[t].optimizeNodes();if(Array.isArray(r))e.splice(t,1,...r);else if(r)e[t]=r;else e.splice(t,1)}return e.length>0?this:undefined}optimizeNames(e,t){const{nodes:r}=this;let n=r.length;while(n--){const s=r[n];if(s.optimizeNames(e,t))continue;subtractNames(e,s.names);r.splice(n,1)}return r.length>0?this:undefined}get names(){return this.nodes.reduce(((e,t)=>addNames(e,t.names)),{})}}class BlockNode extends ParentNode{render(e){return"{"+e._n+super.render(e)+"}"+e._n}}class Root extends ParentNode{}class Else extends BlockNode{}Else.kind="else";class If extends BlockNode{constructor(e,t){super(t);this.condition=e}render(e){let t=`if(${this.condition})`+super.render(e);if(this.else)t+="else "+this.else.render(e);return t}optimizeNodes(){super.optimizeNodes();const e=this.condition;if(e===true)return this.nodes;let t=this.else;if(t){const e=t.optimizeNodes();t=this.else=Array.isArray(e)?new Else(e):e}if(t){if(e===false)return t instanceof If?t:t.nodes;if(this.nodes.length)return this;return new If(not(e),t instanceof If?[t]:t.nodes)}if(e===false||!this.nodes.length)return undefined;return this}optimizeNames(e,t){var r;this.else=(r=this.else)===null||r===void 0?void 0:r.optimizeNames(e,t);if(!(super.optimizeNames(e,t)||this.else))return;this.condition=optimizeExpr(this.condition,e,t);return this}get names(){const e=super.names;addExprNames(e,this.condition);if(this.else)addNames(e,this.else.names);return e}}If.kind="if";class For extends BlockNode{}For.kind="for";class ForLoop extends For{constructor(e){super();this.iteration=e}render(e){return`for(${this.iteration})`+super.render(e)}optimizeNames(e,t){if(!super.optimizeNames(e,t))return;this.iteration=optimizeExpr(this.iteration,e,t);return this}get names(){return addNames(super.names,this.iteration.names)}}class ForRange extends For{constructor(e,t,r,n){super();this.varKind=e;this.name=t;this.from=r;this.to=n}render(e){const t=e.es5?s.varKinds.var:this.varKind;const{name:r,from:n,to:o}=this;return`for(${t} ${r}=${n}; ${r}<${o}; ${r}++)`+super.render(e)}get names(){const e=addExprNames(super.names,this.from);return addExprNames(e,this.to)}}class ForIter extends For{constructor(e,t,r,n){super();this.loop=e;this.varKind=t;this.name=r;this.iterable=n}render(e){return`for(${this.varKind} ${this.name} ${this.loop} ${this.iterable})`+super.render(e)}optimizeNames(e,t){if(!super.optimizeNames(e,t))return;this.iterable=optimizeExpr(this.iterable,e,t);return this}get names(){return addNames(super.names,this.iterable.names)}}class Func extends BlockNode{constructor(e,t,r){super();this.name=e;this.args=t;this.async=r}render(e){const t=this.async?"async ":"";return`${t}function ${this.name}(${this.args})`+super.render(e)}}Func.kind="func";class Return extends ParentNode{render(e){return"return "+super.render(e)}}Return.kind="return";class Try extends BlockNode{render(e){let t="try"+super.render(e);if(this.catch)t+=this.catch.render(e);if(this.finally)t+=this.finally.render(e);return t}optimizeNodes(){var e,t;super.optimizeNodes();(e=this.catch)===null||e===void 0?void 0:e.optimizeNodes();(t=this.finally)===null||t===void 0?void 0:t.optimizeNodes();return this}optimizeNames(e,t){var r,n;super.optimizeNames(e,t);(r=this.catch)===null||r===void 0?void 0:r.optimizeNames(e,t);(n=this.finally)===null||n===void 0?void 0:n.optimizeNames(e,t);return this}get names(){const e=super.names;if(this.catch)addNames(e,this.catch.names);if(this.finally)addNames(e,this.finally.names);return e}}class Catch extends BlockNode{constructor(e){super();this.error=e}render(e){return`catch(${this.error})`+super.render(e)}}Catch.kind="catch";class Finally extends BlockNode{render(e){return"finally"+super.render(e)}}Finally.kind="finally";class CodeGen{constructor(e,t={}){this._values={};this._blockStarts=[];this._constants={};this.opts={...t,_n:t.lines?"\n":""};this._extScope=e;this._scope=new s.Scope({parent:e});this._nodes=[new Root]}toString(){return this._root.render(this.opts)}name(e){return this._scope.name(e)}scopeName(e){return this._extScope.name(e)}scopeValue(e,t){const r=this._extScope.value(e,t);const n=this._values[r.prefix]||(this._values[r.prefix]=new Set);n.add(r);return r}getScopeValue(e,t){return this._extScope.getValue(e,t)}scopeRefs(e){return this._extScope.scopeRefs(e,this._values)}scopeCode(){return this._extScope.scopeCode(this._values)}_def(e,t,r,n){const s=this._scope.toName(t);if(r!==undefined&&n)this._constants[s.str]=r;this._leafNode(new Def(e,s,r));return s}const(e,t,r){return this._def(s.varKinds.const,e,t,r)}let(e,t,r){return this._def(s.varKinds.let,e,t,r)}var(e,t,r){return this._def(s.varKinds.var,e,t,r)}assign(e,t,r){return this._leafNode(new Assign(e,t,r))}add(e,r){return this._leafNode(new AssignOp(e,t.operators.ADD,r))}code(e){if(typeof e=="function")e();else if(e!==n.nil)this._leafNode(new AnyCode(e));return this}object(...e){const t=["{"];for(const[r,s]of e){if(t.length>1)t.push(",");t.push(r);if(r!==s||this.opts.es5){t.push(":");(0,n.addCodeArg)(t,s)}}t.push("}");return new n._Code(t)}if(e,t,r){this._blockNode(new If(e));if(t&&r){this.code(t).else().code(r).endIf()}else if(t){this.code(t).endIf()}else if(r){throw new Error('CodeGen: "else" body without "then" body')}return this}elseIf(e){return this._elseNode(new If(e))}else(){return this._elseNode(new Else)}endIf(){return this._endBlockNode(If,Else)}_for(e,t){this._blockNode(e);if(t)this.code(t).endFor();return this}for(e,t){return this._for(new ForLoop(e),t)}forRange(e,t,r,n,o=(this.opts.es5?s.varKinds.var:s.varKinds.let)){const i=this._scope.toName(e);return this._for(new ForRange(o,i,t,r),(()=>n(i)))}forOf(e,t,r,o=s.varKinds.const){const i=this._scope.toName(e);if(this.opts.es5){const e=t instanceof n.Name?t:this.var("_arr",t);return this.forRange("_i",0,(0,n._)`${e}.length`,(t=>{this.var(i,(0,n._)`${e}[${t}]`);r(i)}))}return this._for(new ForIter("of",o,i,t),(()=>r(i)))}forIn(e,t,r,o=(this.opts.es5?s.varKinds.var:s.varKinds.const)){if(this.opts.ownProperties){return this.forOf(e,(0,n._)`Object.keys(${t})`,r)}const i=this._scope.toName(e);return this._for(new ForIter("in",o,i,t),(()=>r(i)))}endFor(){return this._endBlockNode(For)}label(e){return this._leafNode(new Label(e))}break(e){return this._leafNode(new Break(e))}return(e){const t=new Return;this._blockNode(t);this.code(e);if(t.nodes.length!==1)throw new Error('CodeGen: "return" should have one node');return this._endBlockNode(Return)}try(e,t,r){if(!t&&!r)throw new Error('CodeGen: "try" without "catch" and "finally"');const n=new Try;this._blockNode(n);this.code(e);if(t){const e=this.name("e");this._currNode=n.catch=new Catch(e);t(e)}if(r){this._currNode=n.finally=new Finally;this.code(r)}return this._endBlockNode(Catch,Finally)}throw(e){return this._leafNode(new Throw(e))}block(e,t){this._blockStarts.push(this._nodes.length);if(e)this.code(e).endBlock(t);return this}endBlock(e){const t=this._blockStarts.pop();if(t===undefined)throw new Error("CodeGen: not in self-balancing block");const r=this._nodes.length-t;if(r<0||e!==undefined&&r!==e){throw new Error(`CodeGen: wrong number of nodes: ${r} vs ${e} expected`)}this._nodes.length=t;return this}func(e,t=n.nil,r,s){this._blockNode(new Func(e,t,r));if(s)this.code(s).endFunc();return this}endFunc(){return this._endBlockNode(Func)}optimize(e=1){while(e-- >0){this._root.optimizeNodes();this._root.optimizeNames(this._root.names,this._constants)}}_leafNode(e){this._currNode.nodes.push(e);return this}_blockNode(e){this._currNode.nodes.push(e);this._nodes.push(e)}_endBlockNode(e,t){const r=this._currNode;if(r instanceof e||t&&r instanceof t){this._nodes.pop();return this}throw new Error(`CodeGen: not in block "${t?`${e.kind}/${t.kind}`:e.kind}"`)}_elseNode(e){const t=this._currNode;if(!(t instanceof If)){throw new Error('CodeGen: "else" without "if"')}this._currNode=t.else=e;return this}get _root(){return this._nodes[0]}get _currNode(){const e=this._nodes;return e[e.length-1]}set _currNode(e){const t=this._nodes;t[t.length-1]=e}}t.CodeGen=CodeGen;function addNames(e,t){for(const r in t)e[r]=(e[r]||0)+(t[r]||0);return e}function addExprNames(e,t){return t instanceof n._CodeOrName?addNames(e,t.names):e}function optimizeExpr(e,t,r){if(e instanceof n.Name)return replaceName(e);if(!canOptimize(e))return e;return new n._Code(e._items.reduce(((e,t)=>{if(t instanceof n.Name)t=replaceName(t);if(t instanceof n._Code)e.push(...t._items);else e.push(t);return e}),[]));function replaceName(e){const n=r[e.str];if(n===undefined||t[e.str]!==1)return e;delete t[e.str];return n}function canOptimize(e){return e instanceof n._Code&&e._items.some((e=>e instanceof n.Name&&t[e.str]===1&&r[e.str]!==undefined))}}function subtractNames(e,t){for(const r in t)e[r]=(e[r]||0)-(t[r]||0)}function not(e){return typeof e=="boolean"||typeof e=="number"||e===null?!e:(0,n._)`!${par(e)}`}t.not=not;const a=mappend(t.operators.AND);function and(...e){return e.reduce(a)}t.and=and;const u=mappend(t.operators.OR);function or(...e){return e.reduce(u)}t.or=or;function mappend(e){return(t,r)=>t===n.nil?r:r===n.nil?t:(0,n._)`${par(t)} ${e} ${par(r)}`}function par(e){return e instanceof n.Name?e:(0,n._)`(${e})`}},997:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:true});t.ValueScope=t.ValueScopeName=t.Scope=t.varKinds=t.UsedValueState=void 0;const n=r(621);class ValueError extends Error{constructor(e){super(`CodeGen: "code" for ${e} not defined`);this.value=e.value}}var s;(function(e){e[e["Started"]=0]="Started";e[e["Completed"]=1]="Completed"})(s=t.UsedValueState||(t.UsedValueState={}));t.varKinds={const:new n.Name("const"),let:new n.Name("let"),var:new n.Name("var")};class Scope{constructor({prefixes:e,parent:t}={}){this._names={};this._prefixes=e;this._parent=t}toName(e){return e instanceof n.Name?e:this.name(e)}name(e){return new n.Name(this._newName(e))}_newName(e){const t=this._names[e]||this._nameGroup(e);return`${e}${t.index++}`}_nameGroup(e){var t,r;if(((r=(t=this._parent)===null||t===void 0?void 0:t._prefixes)===null||r===void 0?void 0:r.has(e))||this._prefixes&&!this._prefixes.has(e)){throw new Error(`CodeGen: prefix "${e}" is not allowed in this scope`)}return this._names[e]={prefix:e,index:0}}}t.Scope=Scope;class ValueScopeName extends n.Name{constructor(e,t){super(t);this.prefix=e}setValue(e,{property:t,itemIndex:r}){this.value=e;this.scopePath=(0,n._)`.${new n.Name(t)}[${r}]`}}t.ValueScopeName=ValueScopeName;const o=(0,n._)`\n`;class ValueScope extends Scope{constructor(e){super(e);this._values={};this._scope=e.scope;this.opts={...e,_n:e.lines?o:n.nil}}get(){return this._scope}name(e){return new ValueScopeName(e,this._newName(e))}value(e,t){var r;if(t.ref===undefined)throw new Error("CodeGen: ref must be passed in value");const n=this.toName(e);const{prefix:s}=n;const o=(r=t.key)!==null&&r!==void 0?r:t.ref;let i=this._values[s];if(i){const e=i.get(o);if(e)return e}else{i=this._values[s]=new Map}i.set(o,n);const a=this._scope[s]||(this._scope[s]=[]);const u=a.length;a[u]=t.ref;n.setValue(t,{property:s,itemIndex:u});return n}getValue(e,t){const r=this._values[e];if(!r)return;return r.get(t)}scopeRefs(e,t=this._values){return this._reduceValues(t,(t=>{if(t.scopePath===undefined)throw new Error(`CodeGen: name "${t}" has no value`);return(0,n._)`${e}${t.scopePath}`}))}scopeCode(e=this._values,t,r){return this._reduceValues(e,(e=>{if(e.value===undefined)throw new Error(`CodeGen: name "${e}" has no value`);return e.value.code}),t,r)}_reduceValues(e,r,o={},i){let a=n.nil;for(const u in e){const c=e[u];if(!c)continue;const d=o[u]=o[u]||new Map;c.forEach((e=>{if(d.has(e))return;d.set(e,s.Started);let o=r(e);if(o){const r=this.opts.es5?t.varKinds.var:t.varKinds.const;a=(0,n._)`${a}${r} ${e} = ${o};${this.opts._n}`}else if(o=i===null||i===void 0?void 0:i(e)){a=(0,n._)`${a}${o}${this.opts._n}`}else{throw new ValueError(e)}d.set(e,s.Completed)}))}return a}}t.ValueScope=ValueScope},914:e=>{e.exports=function equal(e,t){if(e===t)return true;if(e&&t&&typeof e=="object"&&typeof t=="object"){if(e.constructor!==t.constructor)return false;var r,n,s;if(Array.isArray(e)){r=e.length;if(r!=t.length)return false;for(n=r;n--!==0;)if(!equal(e[n],t[n]))return false;return true}if(e.constructor===RegExp)return e.source===t.source&&e.flags===t.flags;if(e.valueOf!==Object.prototype.valueOf)return e.valueOf()===t.valueOf();if(e.toString!==Object.prototype.toString)return e.toString()===t.toString();s=Object.keys(e);r=s.length;if(r!==Object.keys(t).length)return false;for(n=r;n--!==0;)if(!Object.prototype.hasOwnProperty.call(t,s[n]))return false;for(n=r;n--!==0;){var o=s[n];if(!equal(e[o],t[o]))return false}return true}return e!==e&&t!==t}}};var t={};function __nccwpck_require__(r){var n=t[r];if(n!==undefined){return n.exports}var s=t[r]={exports:{}};var o=true;try{e[r].call(s.exports,s,s.exports,__nccwpck_require__);o=false}finally{if(o)delete t[r]}return s.exports}if(typeof __nccwpck_require__!=="undefined")__nccwpck_require__.ab=__dirname+"/";var r=__nccwpck_require__(523);module.exports=r})();
