--- packages/next/compiled/ajv-keywords/package.json
@@ -1,1 +1,0 @@
-{"name":"ajv-keywords","main":"index.js","author":"Evgeny Poberezkin","license":"MIT"}
