--- packages/next/compiled/ajv/package.json
@@ -1,1 +1,0 @@
-{"name":"ajv","main":"ajv.js","author":"Evgeny Poberezkin","license":"MIT"}
