+++ packages/next/index.d.ts
@@ -12,7 +12,6 @@
 /// <reference path="./router.d.ts" />
 /// <reference path="./script.d.ts" />
 /// <reference path="./server.d.ts" />
-/// <reference path="./types/compiled.d.ts" />
 
 export { default } from './types'
 export * from './types'
