+++ packages/next/server/config-shared.ts
@@ -1,7 +1,4 @@
 import os from 'os'
-import Ajv from 'next/dist/compiled/ajv'
-import ajvKeywords from 'next/dist/compiled/ajv-keywords'
-import { JSONSchemaType } from 'next/dist/compiled/ajv'
 import type { webpack5 } from 'next/dist/compiled/webpack/webpack'
 import { Header, Redirect, Rewrite } from '../lib/load-custom-routes'
 import {
@@ -9,7 +6,6 @@ import {
   ImageConfigComplete,
   imageConfigDefault,
   RemotePattern,
-  VALID_LOADERS,
 } from '../shared/lib/image-config'
 
 export type ServerRuntime = 'nodejs' | 'experimental-edge' | undefined
@@ -570,613 +566,12 @@ export function isServerRuntime(value?: string): value is ServerRuntime {
   )
 }
 
-const configSchema = {
-  type: 'object',
-  additionalProperties: false,
-  properties: {
-    amp: {
-      additionalProperties: false,
-      properties: {
-        canonicalBase: {
-          minLength: 1,
-          type: 'string',
-        },
-      },
-      type: 'object',
-    },
-    assetPrefix: {
-      minLength: 1,
-      type: 'string',
-    },
-    basePath: {
-      minLength: 1,
-      type: 'string',
-    },
-    cleanDistDir: {
-      type: 'boolean',
-    },
-    compiler: {
-      additionalProperties: false,
-      properties: {
-        emotion: {
-          oneOf: [
-            {
-              type: 'boolean',
-            },
-            {
-              type: 'object',
-              additionalProperties: false,
-              properties: {
-                sourceMap: {
-                  type: 'boolean',
-                },
-                autoLabel: {
-                  type: 'string',
-                  enum: ['always', 'dev-only', 'never'],
-                },
-                labelFormat: {
-                  type: 'string',
-                  minLength: 1,
-                },
-              },
-            },
-          ] as any,
-        },
-        reactRemoveProperties: {
-          oneOf: [
-            {
-              type: 'boolean',
-            },
-            {
-              type: 'object',
-              additionalProperties: false,
-              properties: {
-                properties: {
-                  type: 'array',
-                  items: {
-                    type: 'string',
-                  },
-                },
-              },
-            },
-          ] as any,
-        },
-        relay: {
-          additionalProperties: false,
-          properties: {
-            artifactDirectory: {
-              minLength: 1,
-              type: 'string',
-            },
-            language: {
-              // automatic typing doesn't like enum
-              enum: ['flow', 'typescript'] as any,
-              type: 'string',
-            },
-            src: {
-              minLength: 1,
-              type: 'string',
-            },
-          },
-          type: 'object',
-        },
-        removeConsole: {
-          oneOf: [
-            {
-              type: 'boolean',
-            },
-            {
-              type: 'object',
-              additionalProperties: false,
-              properties: {
-                exclude: {
-                  type: 'array',
-                  items: {
-                    type: 'string',
-                    minLength: 1,
-                  },
-                },
-              },
-            },
-          ] as any,
-        },
-        styledComponents: {
-          oneOf: [
-            {
-              type: 'boolean',
-            },
-            {
-              type: 'object',
-              additionalProperties: false,
-              properties: {
-                displayName: {
-                  type: 'boolean',
-                },
-                topLevelImportPaths: {
-                  type: 'array',
-                  items: {
-                    type: 'string',
-                    minLength: 1,
-                  },
-                },
-                ssr: {
-                  type: 'boolean',
-                },
-                fileName: {
-                  type: 'boolean',
-                },
-                meaninglessFileNames: {
-                  type: 'boolean',
-                },
-                minify: {
-                  type: 'boolean',
-                },
-                transpileTemplateLiterals: {
-                  type: 'boolean',
-                },
-                namespace: {
-                  type: 'string',
-                  minLength: 1,
-                },
-                pure: {
-                  type: 'boolean',
-                },
-                cssProp: {
-                  type: 'boolean',
-                },
-              },
-            },
-          ] as any,
-        },
-      },
-      type: 'object',
-    },
-    compress: {
-      type: 'boolean',
-    },
-    crossOrigin: {
-      oneOf: [
-        false,
-        {
-          enum: ['anonymous', 'use-credentials'],
-          type: 'string',
-        },
-      ], // automatic typing does not like enum
-    } as any,
-    devIndicators: {
-      additionalProperties: false,
-      properties: {
-        buildActivity: {
-          type: 'boolean',
-        },
-        buildActivityPosition: {
-          // automatic typing does not like enum
-          enum: ['bottom-left', 'bottom-right', 'top-left', 'top-right'] as any,
-          type: 'string',
-        },
-      },
-      type: 'object',
-    },
-    distDir: {
-      minLength: 1,
-      type: 'string',
-      nullable: true,
-    },
-    env: {
-      type: 'object',
-    },
-    eslint: {
-      additionalProperties: false,
-      properties: {
-        dirs: {
-          items: {
-            minLength: 1,
-            type: 'string',
-          },
-          type: 'array',
-        },
-        ignoreDuringBuilds: {
-          type: 'boolean',
-        },
-      },
-      type: 'object',
-    },
-    excludeDefaultMomentLocales: {
-      type: 'boolean',
-    },
-    experimental: {
-      additionalProperties: false,
-      properties: {
-        amp: {
-          additionalProperties: false,
-          properties: {
-            optimizer: {
-              type: 'object',
-            },
-            skipValidation: {
-              type: 'boolean',
-            },
-            validator: {
-              type: 'string',
-            },
-          },
-          type: 'object',
-        },
-        appDir: {
-          type: 'boolean',
-        },
-        browsersListForSwc: {
-          type: 'boolean',
-        },
-        cpus: {
-          type: 'number',
-        },
-        craCompat: {
-          type: 'boolean',
-        },
-        disableOptimizedLoading: {
-          type: 'boolean',
-        },
-        disablePostcssPresetEnv: {
-          type: 'boolean',
-        },
-        esmExternals: {
-          type: 'boolean',
-        },
-        externalDir: {
-          type: 'boolean',
-        },
-        forceSwcTransforms: {
-          type: 'boolean',
-        },
-        fullySpecified: {
-          type: 'boolean',
-        },
-        gzipSize: {
-          type: 'boolean',
-        },
-        images: {
-          additionalProperties: false,
-          properties: {
-            allowFutureImage: {
-              type: 'boolean',
-            },
-            remotePatterns: {
-              items: {
-                additionalProperties: false,
-                properties: {
-                  hostname: {
-                    minLength: 1,
-                    type: 'string',
-                  },
-                  pathname: {
-                    minLength: 1,
-                    type: 'string',
-                  },
-                  port: {
-                    minLength: 1,
-                    type: 'string',
-                  },
-                  protocol: {
-                    // automatic typing doesn't like enum
-                    enum: ['http', 'https'] as any,
-                    type: 'string',
-                  },
-                },
-                type: 'object',
-              },
-              type: 'array',
-            },
-            unoptimized: {
-              type: 'boolean',
-            },
-          },
-          type: 'object',
-        },
-        incrementalCacheHandlerPath: {
-          type: 'string',
-        },
-        isrFlushToDisk: {
-          type: 'boolean',
-        },
-        isrMemoryCacheSize: {
-          type: 'number',
-        },
-        largePageDataBytes: {
-          type: 'number',
-        },
-        legacyBrowsers: {
-          type: 'boolean',
-        },
-        manualClientBasePath: {
-          type: 'boolean',
-        },
-        modularizeImports: {
-          type: 'object',
-        },
-        newNextLinkBehavior: {
-          type: 'boolean',
-        },
-        nextScriptWorkers: {
-          type: 'boolean',
-        },
-        optimizeCss: {
-          type: 'boolean',
-        },
-        outputFileTracingRoot: {
-          minLength: 1,
-          type: 'string',
-        },
-        pageEnv: {
-          type: 'boolean',
-        },
-        profiling: {
-          type: 'boolean',
-        },
-        runtime: {
-          // automatic typing doesn't like enum
-          enum: ['experimental-edge', 'nodejs'] as any,
-          type: 'string',
-        },
-        scrollRestoration: {
-          type: 'boolean',
-        },
-        serverComponents: {
-          type: 'boolean',
-        },
-        sharedPool: {
-          type: 'boolean',
-        },
-        swcFileReading: {
-          type: 'boolean',
-        },
-        swcMinify: {
-          type: 'boolean',
-        },
-        swcMinifyDebugOptions: {
-          additionalProperties: false,
-          properties: {
-            compress: {
-              type: 'object',
-            },
-            mangle: {
-              type: 'object',
-            },
-          },
-          type: 'object',
-        },
-        swcPlugins: {
-          type: 'array',
-        },
-        swcTraceProfiling: {
-          type: 'boolean',
-        },
-        urlImports: {
-          items: {
-            type: 'string',
-          },
-          type: 'array',
-        },
-        workerThreads: {
-          type: 'boolean',
-        },
-      },
-      type: 'object',
-    },
-    future: {
-      additionalProperties: false,
-      properties: {},
-      type: 'object',
-    },
-    generateBuildId: {
-      instanceof: 'Function',
-    } as any,
-    generateEtags: {
-      instanceof: 'Function',
-    } as any,
-    headers: {
-      instanceof: 'Function',
-    } as any,
-    httpAgentOptions: {
-      additionalProperties: false,
-      properties: {
-        keepAlive: {
-          type: 'boolean',
-        },
-      },
-      type: 'object',
-    },
-    i18n: {
-      additionalProperties: false,
-      properties: {
-        defaultLocale: {
-          minLength: 1,
-          type: 'string',
-        },
-        domains: {
-          items: {
-            additionalProperties: false,
-            properties: {
-              defaultLocale: {
-                minLength: 1,
-                type: 'string',
-              },
-              domain: {
-                minLength: 1,
-                type: 'string',
-              },
-              http: {
-                type: 'boolean',
-              },
-              locales: {
-                items: {
-                  minLength: 1,
-                  type: 'string',
-                },
-                type: 'array',
-              },
-            },
-            type: 'object',
-          },
-          type: 'array',
-        },
-        localeDetection: {
-          type: 'boolean',
-        },
-        locales: {
-          items: {
-            minLength: 1,
-            type: 'string',
-          },
-          type: 'array',
-        },
-      },
-      type: 'object',
-    },
-    images: {
-      additionalProperties: false,
-      properties: {
-        contentSecurityPolicy: {
-          minLength: 1,
-          type: 'string',
-        },
-        dangerouslyAllowSVG: {
-          type: 'boolean',
-        },
-        deviceSizes: {
-          items: {
-            type: 'number',
-          },
-          minItems: 1,
-          type: 'array',
-        },
-        disableStaticImages: {
-          type: 'boolean',
-        },
-        domains: {
-          items: {
-            type: 'string',
-          },
-          type: 'array',
-        },
-        formats: {
-          items: {
-            enum: ['image/avif', 'image/webp'], // automatic typing does not like enum
-            type: 'string',
-          } as any,
-          type: 'array',
-        },
-        imageSizes: {
-          items: {
-            type: 'number',
-          },
-          minItems: 1,
-          type: 'array',
-        },
-        loader: {
-          // automatic typing does not like enum
-          enum: VALID_LOADERS as any,
-          type: 'string',
-        },
-        minimumCacheTTL: {
-          type: 'number',
-        },
-        path: {
-          minLength: 1,
-          type: 'string',
-        },
-      },
-      type: 'object',
-    },
-    onDemandEntries: {
-      additionalProperties: false,
-      properties: {
-        maxInactiveAge: {
-          type: 'number',
-        },
-        pagesBufferLength: {
-          type: 'number',
-        },
-      },
-      type: 'object',
-    },
-    optimizeFonts: {
-      type: 'boolean',
-    },
-    output: {
-      // automatic typing doesn't like enum
-      enum: ['standalone'] as any,
-      type: 'string',
-    },
-    outputFileTracing: {
-      type: 'boolean',
-    },
-    pageExtensions: {
-      minItems: 1,
-      type: 'array',
-    },
-    poweredByHeader: {
-      type: 'boolean',
-    },
-    productionBrowserSourceMaps: {
-      type: 'boolean',
-    },
-    publicRuntimeConfig: {
-      type: 'object',
-    },
-    reactStrictMode: {
-      type: 'boolean',
-    },
-    redirects: {
-      instanceof: 'Function',
-    } as any,
-    rewrites: {
-      instanceof: 'Function',
-    } as any,
-    sassOptions: {
-      type: 'object',
-    },
-    serverRuntimeConfig: {
-      type: 'object',
-    },
-    staticPageGenerationTimeout: {
-      type: 'number',
-    },
-    swcMinify: {
-      type: 'boolean',
-    },
-    trailingSlash: {
-      type: 'boolean',
-    },
-    typescript: {
-      additionalProperties: false,
-      properties: {
-        ignoreBuildErrors: {
-          type: 'boolean',
-        },
-        tsconfigPath: {
-          minLength: 1,
-          type: 'string',
-        },
-      },
-      type: 'object',
-    },
-    useFileSystemPublicRoutes: {
-      type: 'boolean',
-    },
-    webpack: {
-      instanceof: 'Function',
-    } as any,
-  },
-} as JSONSchemaType<NextConfig>
-
 export function validateConfig(userConfig: NextConfig): {
   errors?: Array<any> | null
 } {
-  const ajv = new Ajv({ allErrors: true })
-  ajvKeywords(ajv, ['instanceof'])
-  ajv.validate(configSchema, userConfig)
+  const configValidator = require('next/dist/next-config-validate.js')
+  configValidator(userConfig)
   return {
-    errors: ajv.errors,
+    errors: configValidator.errors,
   }
 }
