+++ packages/next/tsconfig.json
@@ -8,11 +8,5 @@
     "moduleResolution": "node",
     "jsx": "react"
   },
-  "exclude": [
-    "dist",
-    "./*.d.ts",
-    "future/*.d.ts",
-    "image-types/global.d.ts",
-    "types/compiled.d.ts"
-  ]
+  "exclude": ["dist", "./*.d.ts", "future/*.d.ts", "image-types/global.d.ts"]
 }
