--- packages/next/types/compiled.d.ts
@@ -1,7 +1,0 @@
-declare module 'next/dist/compiled/ajv' {
-  // eslint-disable-next-line
-  export type JSONSchemaType<T> = any
-  function Ajv(..._args: any[]): void
-  export default Ajv
-}
-declare module 'next/dist/compiled/ajv-keywords'
